feat: add policies

+++ app/Http/Controllers/CategoryController.php
@@ -3,76 +3,55 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\CategoryRequest;
+use App\Http\Resources\CategoryResource;
 use App\Models\Category;
 use Inertia\Inertia;
 
 class CategoryController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
-        return Inertia::render('Categories/Index', ['categories' => Category::all()]);
+        return Inertia::render('Categories/Index', [
+            'categories' => CategoryResource::collection(Category::all()),
+        ]);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
         return Inertia::render('Categories/Create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(CategoryRequest $request)
     {
-        Category::create([
-            'name' => $request->input('name'),
-            'description' => $request->input('description'),
-        ]);
-
-        Inertia::flash('message', 'Category created');
-
-        //return redirect()->route('categories.index');
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(Category $category)
-    {
-        return Inertia::render('Categories/Show', ['category' => $category]);
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Category $category)
-    {
-        return Inertia::render('Categories/Edit', ['category' => $category]);
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(CategoryRequest $request, Category $category)
-    {
-        $category->update([
-            'name' => $request->input('name'),
-            'description' => $request->input('description'),
-        ]);
+        Category::create($request->validated());
 
         return redirect()->route('categories.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
+    public function show(Category $category)
+    {
+        return Inertia::render('Categories/Show', [
+            'category' => new CategoryResource($category),
+        ]);
+    }
+
+    public function edit(Category $category)
+    {
+        return Inertia::render('Categories/Edit', [
+            'category' => new CategoryResource($category),
+        ]);
+    }
+
+    public function update(CategoryRequest $request, Category $category)
+    {
+        $category->update($request->validated());
+
+        return redirect()->route('categories.index');
+    }
+
     public function destroy(Category $category)
     {
+        $this->authorize('delete', $category);
         $category->delete();
 
         return redirect()->route('categories.index');
