+++ app/Http/Controllers/CausalController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\CausalStoreRequest;
 use App\Http\Requests\CausalUpdateRequest;
+use App\Http\Resources\CausalResource;
 use App\Models\Causal;
 use Inertia\Inertia;
 
@@ -11,7 +12,9 @@ class CausalController extends Controller
 {
     public function index()
     {
-        return Inertia::render('Causals/Index', ['causals' => Causal::all()]);
+        return Inertia::render('Causals/Index', [
+            'causals' => CausalResource::collection(Causal::all()),
+        ]);
     }
 
     public function create()
@@ -21,34 +24,35 @@ class CausalController extends Controller
 
     public function store(CausalStoreRequest $request)
     {
-        $data = $request->validated();
-
-        Causal::create($data);
+        Causal::create($request->validated());
 
         return redirect()->route('causals.index');
     }
 
     public function show(Causal $causal)
     {
-        return Inertia::render('Causals/Show', ['causal' => $causal]);
+        return Inertia::render('Causals/Show', [
+            'causal' => new CausalResource($causal),
+        ]);
     }
 
     public function edit(Causal $causal)
     {
-        return Inertia::render('Causals/Edit', ['causal' => $causal]);
+        return Inertia::render('Causals/Edit', [
+            'causal' => new CausalResource($causal),
+        ]);
     }
 
     public function update(CausalUpdateRequest $request, Causal $causal)
     {
-        $data = $request->validated();
-
-        $causal->update($data);
+        $causal->update($request->validated());
 
         return redirect()->route('causals.index');
     }
 
     public function destroy(Causal $causal)
     {
+        $this->authorize('delete', $causal);
         $causal->delete();
 
         return redirect()->route('causals.index');
