+++ app/Http/Controllers/ItemController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\ItemStoreRequest;
 use App\Http\Requests\ItemUpdateRequest;
+use App\Http\Resources\CategoryResource;
+use App\Http\Resources\ItemResource;
 use App\Models\Category;
 use App\Models\Item;
 use Inertia\Inertia;
@@ -12,47 +14,50 @@ class ItemController extends Controller
 {
     public function index()
     {
-        return Inertia::render('Items/Index', ['items' => Item::with('category')->get()]);
+        return Inertia::render('Items/Index', [
+            'items' => ItemResource::collection(Item::with('category')->get()),
+        ]);
     }
 
     public function create()
     {
-        return Inertia::render('Items/Create', ['categories' => Category::all()]);
+        return Inertia::render('Items/Create', [
+            'categories' => CategoryResource::collection(Category::all()),
+        ]);
     }
 
     public function store(ItemStoreRequest $request)
     {
-        $data = $request->validated();
-
-        Item::create($data);
+        Item::create($request->validated());
 
         return redirect()->route('items.index');
     }
 
     public function show(Item $item)
     {
-        return Inertia::render('Items/Show', ['item' => $item->load('category', 'suppliers')]);
+        return Inertia::render('Items/Show', [
+            'item' => new ItemResource($item->load('category', 'suppliers')),
+        ]);
     }
 
     public function edit(Item $item)
     {
         return Inertia::render('Items/Edit', [
-            'item' => $item,
-            'categories' => Category::all(),
+            'item'       => new ItemResource($item),
+            'categories' => CategoryResource::collection(Category::all()),
         ]);
     }
 
     public function update(ItemUpdateRequest $request, Item $item)
     {
-        $data = $request->validated();
-
-        $item->update($data);
+        $item->update($request->validated());
 
         return redirect()->route('items.index');
     }
 
     public function destroy(Item $item)
     {
+        $this->authorize('delete', $item);
         $item->delete();
 
         return redirect()->route('items.index');
